Main-2/backend/requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==22.2.0
cachetools==5.3.0
charset-normalizer==3.1.0
click==8.1.3
Flask==2.2.3
Flask-Cors==3.0.10
frozenlist==1.3.3
git-filter-repo
google-api-core
google-auth
google-generativeai
google-ai-generativelanguage
google-cloud-speech
google-cloud-texttospeech
googleapis-common-protos
grpcio
grpcio-status
grpcio-tools
idna==3.4
importlib-metadata==6.3.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
multidict==6.0.4
numpy==1.24.2
pandas==2.0.0
Pillow==9.5.0
proto-plus
protobuf
pyasn1==0.4.8
pyasn1-modules==0.2.8
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
requests==2.28.2
rsa==4.9
six==1.16.0
stability-sdk
tqdm==4.65.0
tzdata==2023.3
urllib3==1.26.15
Werkzeug==2.2.3
yarl==1.8.2
zipp==3.15.0
